Interim commit; initial attempt at using xref
This commit takes the approach that various racket-xp-mode annotations
each have their own xref "backend". Not yet sure if that's smart or
silly.

---

Also not yet confident about how to (a) work efficiently and (b) allow
other non-Racket Mode backends (such as dumbjump or etags) to run,
when we know we can't/won't find anything meaningful, e.g.
xref-find-references generally (except in the limited case of
racket-xp-mode local uses of local definitions). In these general
"can't" cases, we'd like to defer to some other backend that might be
able to find something: We don't want to tell the user "no references
can be found", we want to let some other backend run.

But this seems tricky because each xref method -- such as
xref-backend-identifier-at-point, xref-backend-definitions,
xref-backend-references -- is called independently, and each makes its
own call to xref-find-backend. As a result, each backend's
"activiation function" must do the work of determining whether the
thing at point is something it could handle satisfactorily, or whether
it should punt to some other backend. A naive first
implementation (this commit) ends up calling
xref-backend-identifier-at-point to decide this -- but then xref will
call that one or more times redundantly.

One possibility: Have our backend functions return not a symbol value,
but a list value like (list 'racket <string>), where <string> is from
our current guts of identifier-at-point, possibly propertized with
additional "sub-dispatch" information. Then instead of (eql 'symbol)
the dispatch types are (head 'racket). So the initial determination is
done by the backend function, and (cl-defmethod
xref-back-end-identfier-at-point ((backend (head 'racket)))) just
extracts the string and returns it.

We also supply fallback backend(s) for the case where there are no
annotations and we want to let the user search the "normal" way.

That leaves the fact that (generally) we don't handle
xref-find-references. But our backend function must decide this
upfront. But it isn't told why it's being called -- i.e. what methods
like xref-backend-definitions vs. xref-backend-references will be
called. The least-worst (but bad) idea I can think of so far is to
have the backend function look at the `last-command` variable and see
if it is xref-find-references. O_o ??

Anyway, this commit does NOT yet do any of this, I'm just capturing
notes here, for now, of what I might do in the next commit (in which
case much of this should be turned into comments in the code).