Another take on using the xref API
Here we simply define two xref backends: one each for racket-xp-mode
and racket-repl-mode. The former's xref-backend-find-references will,
if it finds no references, retry while excluding itself, in an attempt
to let other xref backends (such as e.g. dumbjump) run as a fallback.
This is a bit of a hack, as the xref backend API doesn't seem to
anticipate this scenario.